feedback: Use lowest authorization level as default
Presumably (it matches my use-case), anonymous and guest users are less
trained on the use of this software than account-holding users. With
this in mind, this change moves the default authorization level on the
login page to the lowest one.

On a group with anonymous (resp. guest) access:
Before: anonymous (resp. guest) users had to click to select anonymous
(resp. guest) log-in.
After: anonymous (resp. guest) users can log right away, and
password-holding users have to click to log.